fixed declare test, hoist commond deps to workspace
