+++ Cargo.toml
@@ -42,6 +42,7 @@ cairo-felt = "0.8.2"
 anyhow = "1.0.66"
 camino = { version = "1.1.4", features = ["serde1"] }
 include_dir = "0.7.3"
+indoc = "2"
 tempfile = "3.6.0"
 rayon = "1.7"
