+++ crates/cast/Cargo.toml
@@ -25,7 +25,7 @@ which = "4.4.0"
 [dev-dependencies]
 ctor = "0.2.2"
 snapbox = "0.4.11"
-indoc = "2"
+indoc.workspace = true
 url = "2.4.0"
 reqwest = "0.11.18"
 tempfile = "3.3.0"
