+++ crates/cheatable-starknet/src/cheatcodes/declare.rs
@@ -119,9 +119,13 @@ fn felt252_from_hex_string(value: &str) -> Result<Felt252> {
 
 #[cfg(test)]
 mod test {
-    use assert_fs::fixture::PathCopy;
+    use assert_fs::fixture::{FileWriteStr, PathChild, PathCopy};
+    use assert_fs::TempDir;
+    use camino::Utf8PathBuf;
+    use indoc::formatdoc;
     use starknet_api::stark_felt;
     use std::process::Command;
+    use std::str::FromStr;
 
     use super::*;
 
@@ -150,13 +154,39 @@ mod test {
 
     #[test]
     fn class_hash_correct() {
-        let temp = assert_fs::TempDir::new().unwrap();
+        let temp = TempDir::new().unwrap();
         // TODO(#305) change to cheatnet data path
         temp.copy_from(
             "../forge/tests/data/simple_package",
             &["**/*.cairo", "**/*.toml"],
         )
         .unwrap();
+
+        let cheatcodes_path = Utf8PathBuf::from_str("../..")
+            .unwrap()
+            .canonicalize_utf8()
+            .unwrap();
+        println!("{}", cheatcodes_path);
+
+        let manifest_path = temp.child("Scarb.toml");
+        manifest_path
+            .write_str(&formatdoc!(
+                r#"
+                [package]
+                name = "simple_package"
+                version = "0.1.0"
+        
+                [[target.starknet-contract]]
+                sierra = true
+                casm = true
+        
+                [dependencies]
+                starknet = "2.1.0-rc2"
+                cheatcodes = {{ path = "{}" }}
+                "#,
+                cheatcodes_path
+            ))
+            .unwrap();
 
         Command::new("scarb")
             .current_dir(&temp)
