+++ crates/forge/Cargo.toml
@@ -10,6 +10,7 @@ anyhow.workspace = true
 blockifier.workspace = true
 camino.workspace = true
 include_dir.workspace = true
+indoc.workspace = true
 starknet_api.workspace = true
 tempfile.workspace = true
 rayon.workspace = true
@@ -40,7 +41,6 @@ assert_fs = "1.0.13"
 snapbox = "0.4.11"
 starknet = "0.4.0"
 once_cell = "1.18.0"
-indoc = "2.0.1"
 cairo-vm = "0.8.1"
 ark-ff = "0.4.0-alpha.7"
 ark-std = "0.4.0"
