+++ crates/forge/tests/data/simple_package/Scarb.toml
@@ -4,7 +4,6 @@ version = "0.1.0"
 
 [dependencies]
 starknet = "2.1.0-rc2"
-#cheatcodes = { git = "https://github.com/foundry-rs/starknet-foundry", branch = "68-export-cheatcodes-as-scarb-package"}
 cheatcodes = { path = "../../../../.." }
 
 [[target.starknet-contract]]
