+++ crates/forge/tests/e2e/running.rs
@@ -1,47 +1,10 @@
-use assert_fs::fixture::{FileWriteStr, PathChild, PathCopy};
-use assert_fs::TempDir;
+use assert_fs::fixture::{FileWriteStr, PathChild};
 use camino::Utf8PathBuf;
 use indoc::{formatdoc, indoc};
 
 use crate::e2e::common::runner::runner;
+use forge::scarb::tests::setup_package;
 use std::str::FromStr;
-
-fn setup_package(package_name: &str) -> TempDir {
-    let temp = TempDir::new().unwrap();
-    temp.copy_from(
-        format!("tests/data/{package_name}"),
-        &["**/*.cairo", "**/*.toml"],
-    )
-    .unwrap();
-
-    let cheatcodes_path = Utf8PathBuf::from_str("../..")
-        .unwrap()
-        .canonicalize_utf8()
-        .unwrap();
-
-    let manifest_path = temp.child("Scarb.toml");
-    manifest_path
-        .write_str(&formatdoc!(
-            r#"
-            [package]
-            name = "{}"
-            version = "0.1.0"
-
-            [[target.starknet-contract]]
-            sierra = true
-            casm = true
-
-            [dependencies]
-            starknet = "2.1.0-rc2"
-            cheatcodes = {{ path = "{}" }}
-            "#,
-            package_name,
-            cheatcodes_path
-        ))
-        .unwrap();
-
-    temp
-}
 
 #[test]
 fn simple_package() {
